cloud-functions/scrape-tickers/requirements.txt
pandas==2.2.2
google-cloud-storage==2.17.0
google-cloud-secret-manager==2.17.0